Add missing trace dependency to kubemark
